- fix warnings - missed an ecore function rename
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@31404 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/ewl/src/lib/ewl_text.c
@@ -2876,7 +2876,7 @@ ewl_text_cb_format(Ewl_Text_Fmt_Node *node, Ewl_Text *t, unsigned int byte_idx)
 	tmp = *(ptr + node->byte_len);
 	if (strlen(ptr) < node->byte_len)
 		DWARNING("Byte length of node %u overruns actual text %d",
-				node->byte_len, strlen(ptr));
+				node->byte_len, (int)strlen(ptr));
 	*(ptr + node->byte_len) = '\0';
 
 	ewl_text_plaintext_parse(t->textblock, ptr);
